Feature: #161242048 Modularizing the api
