+++ app/api/v1/views/products_views.py
@@ -6,6 +6,7 @@ from functools import wraps
 # local imports
 from ..models.products import Product as ProductClass
 from ..utils.dto import ProductsDto, product_parser, update_product_parser
+from ..utils.decorators import token_required
 
 api = ProductsDto.api
 products = ProductsDto.products
@@ -13,25 +14,6 @@ post_products = ProductsDto.post_products
 
 product = ProductClass()
 
-def token_required(f):
-    @wraps(f)
-    def decorated(*args, **kwargs):
-
-        token = None
-
-        if 'x-access-token' in request.headers:
-            token = request.headers['x-access-token']
-
-        if not token:
-            return {'message' : 'Token is missing.'}, 401
-
-        if token != 'mytoken':
-            return {'message' : 'Your token is wrong, wrong, wrong!!!'}, 401
-
-        print('TOKEN: {}'.format(token))
-        return f(*args, **kwargs)
-
-    return decorated
 
 @api.route("/products")
 class ProductList(Resource):
@@ -40,7 +22,7 @@ class ProductList(Resource):
     @api.expect(post_products)
     @api.doc('creates an product', security='apikey')
     @api.response(201, "Created")
-    
+    @token_required
     def post(self):
         """Creates a new Product."""
         args = product_parser.parse_args()
@@ -49,7 +31,7 @@ class ProductList(Resource):
     @api.doc("list_products", security='apikey')
     @api.response(404, "Products Not Found")
     @api.marshal_list_with(products, envelope="products")
-    
+    @token_required
     def get(self):
         """List all Products"""
         return product.get_all()
@@ -62,7 +44,7 @@ class Product(Resource):
 
     @api.marshal_with(products)
     @api.doc('get one product', security='apikey')
-    
+    @token_required
     def get(self, productId):
         """Displays a single Product."""
         return product.get_one(productId)
@@ -70,7 +52,7 @@ class Product(Resource):
     @api.marshal_with(products)
     @api.doc('updates an product', security='apikey')
     @api.expect(post_products)
-    
+    @token_required
     def put(self, productId):
         """Updates a single Product."""
         args = update_product_parser.parse_args()
@@ -79,7 +61,7 @@ class Product(Resource):
     @api.marshal_with(products)
     @api.doc('deletes an product', security='apikey')
     @api.response(204, 'Product Deleted')
-    
+    @token_required
     def delete(self, productId):
         """Deletes a single Product."""
         product.delete_product(productId)
